cleanup morse generation

+++ morse.h
@@ -4,6 +4,6 @@
 #include <wchar.h>
 #include <wctype.h>
 
-int encode_morse_char(int16_t *samples,wint_t c);
+int encode_morse_char(float *samples,wint_t c);
 int init_morse(float const speed,float const pitch,float const level,float const samprate);
 #endif
